Move dummy components
